See #70: Execute handler on original event in listen-live.
Currently, the original event is discarded, and the event handler is run
on a new goog.events.Event with an appropriate currentTarget property.
This makes .preventDefault and .stopPropagation impossible.

With this commit, listen-live will execute the handler on the copy, so
that the handler can read the desired currentTarget. Then it will read
the copy's defaultPrevented and propagationStopped properties to decide
whether to call preventDefault and stopPropagation on the original
"real" event.